perbaikan untuk delete dan view

+++ Backend/db.php
@@ -31,7 +31,7 @@ class Database
     // Tampilkan semua pemilih
     function viewPemilih()
     {
-        $query = $this->connect->query("SELECT * FROM pengguna");
+        $query = $this->connect->query("SELECT * FROM pengguna WHERE role = 'user'");
         return $query->fetch_all(MYSQLI_ASSOC);
     }
 
@@ -103,7 +103,7 @@ class Database
     // Hapus semua pemilih
     public function deleteAllPemilih()
     {
-        $stmt = $this->connect->prepare("DELETE FROM pengguna");
+        $stmt = $this->connect->prepare("DELETE FROM pengguna WHERE role = 'user'");
         return $stmt->execute();
     }
